allow tmpdir to be specified in config file
Both CheckM and GTDB-Tk can benefit from writing files to a temporary directory to reduce memory overhead. The location of this directory was hard-coded to `/scratch`, which assumes usage of HPC and access to directory. Allowing the location to be specified in the config file circumvents issues when `/scratch` is not available on HPC or when snakemake is run locally. The config file, snakefile, and documentation have been updated to reflect this change.