Allow fasta chunks to be specified
A new option is implemented which allows the user to change the number of fasta chunks created before alignments. The previous version hard-coded 4 chunks (DIAMOND) or 2 chunks (minimap2). These values are optimal for a fasta with ~2.5 million HiFi reads, but may not be optimal for smaller fasta files, which may benefit from fewer chunks (2 or 1). Conversely, larger files may benefit from a bump in chunk number. However, the upper chunk limit is 9, and over-chunking will dramatically slow the workflows. Use of defaults is strongly recommended for most datasets.