bug fix for large refs
Given a reference longer than 4Gb, minimap2 is unable to see all the sequences and thus can't produce a correct SAM header. This throws an error in the `MinimapToBam` rule with the sorting step. This fix avoids the issue by splitting this rule into two new rules: `MinimapIndex` and `MinimapToBam`. indexing the reference prior to alignment and subsequently running minimap2 with the `--split-prefix` option.